backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
praw==7.8.1
prawcore==2.4.0
pydantic==2.11.5
pydantic_core==2.33.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
psycopg2-binary